src/villajets/pyproject.toml
[project]
name = "villajets"
version = "0.1.0"
description = ""
authors = [
    {name = "Jose Garcia Escandón",email = "jose.g@covergenius.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.2.1,<6.0.0)",
    "django-unfold (>=0.57.0,<0.58.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "faker (>=37.3.0,<38.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)"
]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
